evidence/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
psycopg2-binary==2.9.9
pydantic==2.4.2
